+++ src/main/java/uk/co/shockwaveInteractive/objects/enchantment/AtomRipperEnchantment.java
@@ -1,0 +1,53 @@
+package uk.co.shockwaveinteractive.objects.enchantment;
+
+import net.minecraft.enchantment.Enchantment;
+import net.minecraft.enchantment.EnchantmentType;
+import net.minecraft.entity.CreatureAttribute;
+import net.minecraft.entity.Entity;
+import net.minecraft.entity.LivingEntity;
+import net.minecraft.inventory.EquipmentSlotType;
+import net.minecraft.item.ItemStack;
+import net.minecraft.potion.Effect;
+import net.minecraft.potion.EffectInstance;
+import net.minecraft.potion.Effects;
+
+public class AtomRipperEnchantment extends Enchantment {
+
+    public AtomRipperEnchantment() {
+        super(Rarity.VERY_RARE, EnchantmentType.WEAPON, new EquipmentSlotType[]{EquipmentSlotType.MAINHAND, EquipmentSlotType.OFFHAND});
+    }
+
+    @Override
+    public int getMaxEnchantability(int enchantmentLevel) {
+        return 25;
+    }
+
+    @Override
+    public int getMaxLevel() {
+        return 1;
+    }
+
+    @Override
+    public boolean canApplyAtEnchantingTable(ItemStack stack) {
+        return false;
+    }
+
+    @Override
+    public boolean isAllowedOnBooks() {
+        return true;
+    }
+
+    @Override
+    public boolean isTreasureEnchantment() {
+        return true;
+    }
+
+    @Override
+    public void onEntityDamaged(LivingEntity user, Entity target, int level) {
+        if(target.isLiving() && ((LivingEntity)target).getCreatureAttribute() == CreatureAttribute.UNDEAD)
+        {
+            target.setFire(5);
+            user.addPotionEffect(new EffectInstance(Effects.REGENERATION, 100));
+        }
+    }
+}